Week5/user-service/requirements.txt
Flask==2.1.1
psycopg2-binary==2.9.3
werkzeug==2.0.3
gunicorn==20.1.0

